fix(offline): SSR to SSG.
Because Svelte was set to SRR a server was required to run the app.
To make it run without this server, I had to convert it into a SSG.

There were also fixes for file paths to be absolute instead of relative.
I set it to relative by accident. "icons/icon.jpg" to "/icons/icon.jpg"
the leading / made it absolute.