fix(modal): #5 duplicate activities
in #5 the issue is duplicate activities. Bad code was written which
results in two issues. 1. Duplicate entries were added into the db.

The primary key is the id of the activity, not the name. Attempting to
update the table to remove id and have the name be the primary key
resulted in an error. It is an unsupported operation in DexieJS.

The each block in the modal requires a key.The name of the activity is
the chosen key. However, due to duplicate entries in the db the modal
throws an error and doesn't appear as each key must be unique.

2. Duplicate entries were not added into the db but simply rendered
multiple times.

After adding a simple guard to prevent duplicate activity names the
duplicate re-renders never happens.

Notes on refactor: All UI variables now are delcared with $state and
derived with $derived. This allows UI elements to update on demand. This
means that when adding a new activity, the new option appears instantly.
Previously, the modal needs to be closed and re-opened to see the new
item.