feat: added activity and project management features to SelectModal and ProjectModal.
SelectModal (#86, #87) Users are now able to delete, edit, activities
and tasks in the SelectModal.

The way this change was made was by separating tasks and activites and
have it be combined into one inside SelectModal.

ProjectModal Users are now able to delete, edit project name and color
in the ProjectModal.